refactor: move sweeper and pulses to seperate modules
