+++ src/qibolab/_core/instruments/keysight/__init__.py
@@ -1,0 +1,5 @@
+from . import qcs
+from .qcs import *
+
+__all__ = []
+__all__ += qcs.__all__
