--- src/qibolab/_core/instruments/keysight_qcs.py
@@ -1,384 +1,0 @@
-"""Qibolab driver for Keysight QCS instrument set."""
-
-from collections import defaultdict
-from functools import reduce
-from typing import ClassVar, Optional, Union
-
-import keysight.qcs as qcs  # pylint: disable=E0401
-import numpy as np
-
-from qibolab._core.components import AcquisitionChannel, Config, DcChannel, IqChannel
-from qibolab._core.execution_parameters import (
-    AcquisitionType,
-    AveragingMode,
-    ExecutionParameters,
-)
-from qibolab._core.identifier import ChannelId, Result
-from qibolab._core.instruments.abstract import Controller
-from qibolab._core.pulses import Drag, Envelope, Gaussian, PulseId, Rectangular
-from qibolab._core.sequence import InputOps, PulseSequence
-from qibolab._core.sweeper import ParallelSweepers, Parameter
-from qibolab._core.unrolling import Bounds
-
-NS_TO_S = 1e-9
-
-BOUNDS = Bounds(
-    waveforms=1,
-    readout=1,
-    instructions=1,
-)
-
-__all__ = ["KeysightQCS"]
-
-
-def generate_qcs_envelope(shape: Envelope) -> qcs.Envelope:
-    """Converts a Qibolab pulse envelope to a QCS Envelope object."""
-    if isinstance(shape, Rectangular):
-        return qcs.ConstantEnvelope()
-
-    elif isinstance(shape, (Gaussian, Drag)):
-        return qcs.GaussianEnvelope(shape.rel_sigma)
-
-    else:
-        # TODO: Rework this code to support other Qibolab pulse envelopes
-        # raw_envelope = shape.i(num_samples) + 1j * shape.q(num_samples)
-        # return qcs.ArbitraryEnvelope(
-        #    times=np.linspace(0, 1, num_samples), amplitudes=raw_envelope
-        # )
-        raise Exception("Envelope not supported")
-
-
-def generate_qcs_rfwaveform(
-    duration: Union[float, qcs.Scalar],
-    envelope: Envelope,
-    amplitude: Union[float, qcs.Scalar],
-    frequency: Union[float, qcs.Scalar],
-    phase: Union[float, qcs.Scalar],
-) -> qcs.RFWaveform:
-
-    return qcs.RFWaveform(
-        duration=duration,
-        envelope=generate_qcs_envelope(envelope),
-        amplitude=amplitude,
-        rf_frequency=frequency,
-        instantaneous_phase=phase,
-    )
-
-
-def process_sweepers(
-    sweepers: list[ParallelSweepers], probe_channel_ids: set[ChannelId]
-):
-    hardware_sweepers = []
-    software_sweepers = []
-    hw_sweeper_order = []
-    sw_sweeper_order = []
-
-    # Mapper for pulses that are controlled by a sweeper and the parameter to be swept
-    sweeper_pulse_map: defaultdict[PulseId, dict[str, qcs.Scalar]] = defaultdict(dict)
-    # Mapper for channels with frequency controlled by a sweeper
-    sweeper_channel_map: dict[ChannelId, qcs.Scalar] = {}
-
-    for idx, parallel_sweeper in enumerate(sweepers):
-        sweep_values: list[qcs.Array] = []
-        sweep_variables: list[qcs.Variable] = []
-        # Currently nested hardware sweeping is not supported
-        hardware_sweeping = len(hardware_sweepers) == 0
-
-        for idx2, sweeper in enumerate(parallel_sweeper):
-            qcs_variable = qcs.Scalar(
-                name=f"V{idx}_{idx2}", value=sweeper.values[0], dtype=float
-            )
-
-            if sweeper.parameter is Parameter.frequency:
-                sweeper_channel_map.update(
-                    {channel_id: qcs_variable for channel_id in sweeper.channels}
-                )
-                # Readout frequency is not supported with hardware sweeping
-                if not probe_channel_ids.isdisjoint(sweeper.channels):
-                    hardware_sweeping = False
-            elif sweeper.parameter in [
-                Parameter.amplitude,
-                Parameter.duration,
-                Parameter.relative_phase,
-            ]:
-                # Duration is not supported with hardware sweeping
-                if sweeper.parameter is Parameter.duration:
-                    hardware_sweeping = False
-
-                for pulse in sweeper.pulses:
-                    sweeper_pulse_map[pulse.id][sweeper.parameter.name] = qcs_variable
-            else:
-                raise ValueError(
-                    "Sweeper parameter not supported", sweeper.parameter.name
-                )
-
-            sweep_variables.append(qcs_variable)
-            sweep_values.append(
-                qcs.Array(
-                    name=f"A{idx}_{idx2}",
-                    value=(
-                        sweeper.values * NS_TO_S
-                        if sweeper.parameter is Parameter.duration
-                        else sweeper.values
-                    ),
-                    dtype=float,
-                )
-            )
-        if hardware_sweeping:
-            hardware_sweepers.append((sweep_values, sweep_variables))
-            hw_sweeper_order.append(idx)
-        else:
-            software_sweepers.append((sweep_values, sweep_variables))
-            sw_sweeper_order.append(idx)
-
-    sweeper_swaps_required = [
-        (original_index, shifted_index)
-        for shifted_index, original_index in enumerate(
-            hw_sweeper_order + sw_sweeper_order
-        )
-        if original_index != shifted_index
-    ]
-    return (
-        hardware_sweepers,
-        software_sweepers,
-        sweeper_swaps_required,
-        sweeper_channel_map,
-        sweeper_pulse_map,
-    )
-
-
-class KeysightQCS(Controller):
-    """Driver for interacting with QCS controller server."""
-
-    bounds: str = "qcs/bounds"
-
-    # Map of QCS virtual channels to QCS physical channels
-    qcs_channel_map: qcs.ChannelMapper
-    # Map of Qibolab channel IDs to QCS virtual channels
-    virtual_channel_map: dict[ChannelId, qcs.Channels]
-    # Map of QCS virtual acquisition channels to QCS state classifiers
-    classifier_map: Optional[dict[qcs.Channels, qcs.MinimumDistanceClassifier]] = None
-    sampling_rate: ClassVar[float] = (
-        qcs.SAMPLE_RATES[qcs.InstrumentEnum.M5300AWG] * NS_TO_S
-    )
-
-    def connect(self):
-        self.backend = qcs.HclBackend(self.qcs_channel_map, hw_demod=True)
-        self.backend.is_system_ready()
-
-    def create_program(
-        self,
-        sequence: PulseSequence,
-        configs: dict[str, Config],
-        sweepers: list[ParallelSweepers],
-        num_shots: int,
-    ) -> tuple[qcs.Program, list[tuple[int, int]]]:
-        program = qcs.Program()
-
-        # SWEEPER MANAGEMENT
-        probe_channel_ids = {
-            chan.probe
-            for chan in self.channels.values()
-            if isinstance(chan, AcquisitionChannel)
-        }
-
-        (
-            hardware_sweepers,
-            software_sweepers,
-            sweeper_swaps_required,
-            sweeper_channel_map,
-            sweeper_pulse_map,
-        ) = process_sweepers(sweepers, probe_channel_ids)
-        # Here we are telling the program to run hardware sweepers first, then software sweepers
-        # It is essential that we match the original sweeper order to the modified sweeper order
-        # to reconcile the results at the end
-        sweep = lambda program, sweepers: program.sweep(*sweepers)
-        program = reduce(
-            sweep,
-            software_sweepers,
-            reduce(sweep, hardware_sweepers, program).n_shots(num_shots),
-        )
-
-        # WAVEFORM COMPILATION
-        # Iterate over channels and convert qubit pulses to QCS waveforms
-        for channel_id in sequence.channels:
-            channel = self.channels[channel_id]
-            virtual_channel = self.virtual_channel_map[channel_id]
-
-            if isinstance(channel, AcquisitionChannel):
-                probe_channel_id = channel.probe
-                probe_virtual_channel = self.virtual_channel_map[probe_channel_id]
-                if probe_channel_id in sweeper_channel_map:
-                    frequency = sweeper_channel_map[probe_channel_id]
-                else:
-                    frequency = configs[probe_channel_id].frequency
-                for pulse in sequence.channel(channel_id):
-                    sweep_param_map = sweeper_pulse_map.get(pulse.id, {})
-
-                    if pulse.kind == "delay":
-                        qcs_pulse = qcs.Delay(
-                            sweep_param_map.get("duration", pulse.duration * NS_TO_S)
-                        )
-                        program.add_waveform(qcs_pulse, virtual_channel)
-                        program.add_waveform(qcs_pulse, probe_virtual_channel)
-
-                    elif pulse.kind == "acquisition":
-                        duration = sweep_param_map.get(
-                            "duration", pulse.duration * NS_TO_S
-                        )
-                        program.add_acquisition(duration, virtual_channel)
-
-                    elif pulse.kind == "readout":
-                        sweep_param_map = sweeper_pulse_map.get(pulse.probe.id, {})
-                        qcs_pulse = generate_qcs_rfwaveform(
-                            duration=sweep_param_map.get(
-                                "duration", pulse.probe.duration * NS_TO_S
-                            ),
-                            envelope=pulse.probe.envelope,
-                            amplitude=sweep_param_map.get(
-                                "amplitude", pulse.probe.amplitude
-                            ),
-                            frequency=frequency,
-                            phase=sweep_param_map.get(
-                                "relative_phase", pulse.probe.relative_phase
-                            ),
-                        )
-                        integration_filter = qcs.IntegrationFilter(qcs_pulse)
-                        program.add_waveform(qcs_pulse, probe_virtual_channel)
-                        program.add_acquisition(integration_filter, virtual_channel)
-
-            elif isinstance(channel, IqChannel):
-
-                pulses = []
-                if channel_id in sweeper_channel_map:
-                    frequency = sweeper_channel_map[channel_id]
-                else:
-                    frequency = configs[channel_id].frequency
-
-                for pulse in sequence.channel(channel_id):
-                    sweep_param_map = sweeper_pulse_map.get(pulse.id, {})
-
-                    if pulse.kind == "delay":
-                        qcs_pulse = qcs.Delay(
-                            sweep_param_map.get("duration", pulse.duration * NS_TO_S)
-                        )
-                    elif pulse.kind == "virtualz":
-                        qcs_pulse = qcs.PhaseIncrement(
-                            phase=sweep_param_map.get("relative_phase", pulse.phase)
-                        )
-                    elif pulse.kind == "pulse":
-                        qcs_pulse = generate_qcs_rfwaveform(
-                            duration=sweep_param_map.get(
-                                "duration", pulse.duration * NS_TO_S
-                            ),
-                            envelope=pulse.envelope,
-                            amplitude=sweep_param_map.get("amplitude", pulse.amplitude),
-                            frequency=frequency,
-                            phase=sweep_param_map.get(
-                                "relative_phase", pulse.relative_phase
-                            ),
-                        )
-                        if pulse.envelope.kind == "drag":
-                            qcs_pulse = qcs_pulse.drag(coeff=pulse.envelope.beta)
-                    else:
-                        raise ValueError("Unrecognized pulse type", pulse.kind)
-
-                    pulses.append(qcs_pulse)
-
-                program.add_waveform(pulses, virtual_channel)
-
-            elif isinstance(channel, DcChannel):
-
-                pulses = []
-                sweep_param_map = sweeper_pulse_map.get(pulse.id, {})
-
-                for pulse in sequence.channel(channel_id):
-                    if pulse.kind == "delay":
-                        qcs_pulse = qcs.Delay(
-                            sweep_param_map.get("duration", pulse.duration * NS_TO_S)
-                        )
-                    elif pulse.kind == "pulse":
-                        qcs_pulse = qcs.DCWaveform(
-                            duration=sweep_param_map.get(
-                                "duration", pulse.duration * NS_TO_S
-                            ),
-                            envelope=generate_qcs_envelope(pulse.envelope),
-                            amplitude=sweep_param_map.get("amplitude", pulse.amplitude),
-                        )
-                    else:
-                        raise ValueError("Unrecognized pulse type", pulse.kind)
-                    pulses.append(qcs_pulse)
-
-                program.add_waveform(pulses, virtual_channel)
-        return program, sweeper_swaps_required
-
-    def play(
-        self,
-        configs: dict[str, Config],
-        sequences: list[PulseSequence],
-        options: ExecutionParameters,
-        sweepers: list[ParallelSweepers],
-    ) -> dict[int, Result]:
-
-        if options.relaxation_time is not None:
-            self.backend._init_time = int(options.relaxation_time)
-
-        ret: dict[PulseId, np.ndarray] = {}
-        for sequence in sequences:
-            program, sweeper_swaps_required = self.create_program(
-                sequence.align_to_delays(), configs, sweepers, options.nshots
-            )
-            results = self.backend.apply(program).results
-            acquisition_map: defaultdict[qcs.Channels, list[InputOps]] = defaultdict(
-                list
-            )
-
-            for channel_id, input_op in sequence.acquisitions:
-                channel = self.virtual_channel_map[channel_id]
-                acquisition_map[channel].append(input_op)
-
-            averaging = options.averaging_mode is not AveragingMode.SINGLESHOT
-            for channel, input_ops in acquisition_map.items():
-                if options.acquisition_type is AcquisitionType.RAW:
-                    raw = results.get_trace(channel, averaging)
-                elif options.acquisition_type is AcquisitionType.INTEGRATION:
-                    raw = results.get_iq(channel, averaging)
-                elif options.acquisition_type is AcquisitionType.DISCRIMINATION:
-                    classifier = self.classifier_map[channel]
-                    raw = results.get_classified(channel, averaging, classifier)
-                else:
-                    raise ValueError("Acquisition type unrecognized")
-
-                if len(sweeper_swaps_required) > 0:
-                    swap = lambda array, pair_indices: np.swapaxes(array, *pair_indices)
-                    for key, value in raw.items():
-                        raw[key] = reduce(swap, sweeper_swaps_required, value)
-
-                for result, input_op in zip(raw.values(), input_ops):
-                    # For single shot, qibolab expects result format (nshots, ...)
-                    # QCS returns (..., nshots), so we need to shuffle the arrays
-                    if (
-                        options.averaging_mode is AveragingMode.SINGLESHOT
-                        and len(sweepers) > 0
-                    ):
-                        tmp = np.zeros(options.results_shape(sweepers))
-                        if options.acquisition_type is AcquisitionType.INTEGRATION:
-                            for k in range(options.nshots):
-                                tmp[k, ..., 0] = np.real(result[..., k])
-                                tmp[k, ..., 1] = np.imag(result[..., k])
-                        else:
-                            for k in range(options.nshots):
-                                tmp[k, ...] = result[..., k]
-                        result = tmp
-
-                    elif options.acquisition_type is AcquisitionType.INTEGRATION:
-                        tmp = np.zeros(result.shape + (2,))
-                        tmp[..., 0] = np.real(result)
-                        tmp[..., 1] = np.imag(result)
-                        result = tmp
-                    ret[input_op.id] = result
-
-        return ret
-
-    def disconnect(self):
-        pass
