+++ src/qibolab/instruments/keysight_qcs.py
@@ -1,7 +1,7 @@
 """Driver for Keysight Quantum Control System (QCS)."""
 
-from qibolab._core.instruments import keysight_qcs
-from qibolab._core.instruments.keysight_qcs import *  # noqa: F403
+from qibolab._core.instruments.keysight import qcs
+from qibolab._core.instruments.keysight.qcs import *  # noqa: F403
 
 __all__ = []
-__all__ += keysight_qcs.__all__
+__all__ += qcs.__all__
